+++ android/src/newarch/java/io/sentry/react/RNSentryModule.java
@@ -158,4 +158,14 @@ public class RNSentryModule extends NativeRNSentrySpec {
         // Not used on Android
         return null;
     }
+
+    @Override
+    public void captureReplay(boolean isHardCrash, Promise promise) {
+        this.impl.captureReplay(isHardCrash, promise);
+    }
+
+    @Override
+    public String getCurrentReplayId() {
+        return this.impl.getCurrentReplayId();
+    }
 }