+++ src/js/client.ts
@@ -17,6 +17,8 @@ import { Alert } from 'react-native';
 import { createIntegration } from './integrations/factory';
 import { defaultSdkInfo } from './integrations/sdkinfo';
 import type { ReactNativeClientOptions } from './options';
+import type { mobileReplayIntegration } from './replay/mobilereplay';
+import { MOBILE_REPLAY_INTEGRATION_NAME } from './replay/mobilereplay';
 import type { ReactNativeTracing } from './tracing';
 import { createUserFeedbackEnvelope, items } from './utils/envelope';
 import { ignoreRequireCycleLogs } from './utils/ignorerequirecyclelogs';
@@ -130,8 +132,8 @@ export class ReactNativeClient extends BaseClient<ReactNativeClientOptions> {
    * @inheritDoc
    */
   public init(): void {
-    this._initNativeSdk();
     super.init();
+    this._initNativeSdk();
   }
 
   /**
@@ -154,7 +156,14 @@ export class ReactNativeClient extends BaseClient<ReactNativeClientOptions> {
    * Starts native client with dsn and options
    */
   private _initNativeSdk(): void {
-    NATIVE.initNativeSdk(this._options)
+    NATIVE.initNativeSdk({
+      ...this._options,
+      mobileReplayOptions:
+        this._integrations[MOBILE_REPLAY_INTEGRATION_NAME] &&
+        'options' in this._integrations[MOBILE_REPLAY_INTEGRATION_NAME]
+          ? (this._integrations[MOBILE_REPLAY_INTEGRATION_NAME] as ReturnType<typeof mobileReplayIntegration>).options
+          : undefined,
+    })
       .then(
         (result: boolean) => {
           return result;