+++ src/js/integrations/exports.ts
@@ -12,6 +12,7 @@ export { screenshotIntegration } from './screenshot';
 export { viewHierarchyIntegration } from './viewhierarchy';
 export { expoContextIntegration } from './expocontext';
 export { spotlightIntegration } from './spotlight';
+export { mobileReplayIntegration } from '../replay/mobilereplay';
 export { appStartIntegration } from '../tracing/integrations/appStart';
 
 export {
@@ -25,4 +26,5 @@ export {
   inboundFiltersIntegration,
   linkedErrorsIntegration as browserLinkedErrorsIntegration,
   rewriteFramesIntegration,
+  replayIntegration as browserReplayIntegration,
 } from '@sentry/react';