+++ test/touchevents.test.tsx
@@ -5,7 +5,7 @@ import * as core from '@sentry/core';
 import type { SeverityLevel } from '@sentry/types';
 
 import { TouchEventBoundary } from '../src/js/touchevents';
-import { getDefaultTestClientOptions,TestClient } from './mocks/client';
+import { getDefaultTestClientOptions, TestClient } from './mocks/client';
 
 describe('TouchEventBoundary._onTouchStart', () => {
   let addBreadcrumb: jest.SpyInstance;
@@ -101,7 +101,7 @@ describe('TouchEventBoundary._onTouchStart', () => {
     expect(addBreadcrumb).toBeCalledWith({
       category: defaultProps.breadcrumbCategory,
       data: {
-        componentTree: ['View', 'Connect(View)', 'LABEL!'],
+        path: [{ name: 'View' }, { name: 'Connect(View)' }, { label: 'LABEL!' }],
       },
       level: 'info' as SeverityLevel,
       message: 'Touch event within element: LABEL!',
@@ -160,15 +160,15 @@ describe('TouchEventBoundary._onTouchStart', () => {
     expect(addBreadcrumb).toBeCalledWith({
       category: defaultProps.breadcrumbCategory,
       data: {
-        componentTree: ['Styled(View2)', 'Styled(View)'],
+        path: [{ name: 'Styled(View)' }],
       },
       level: 'info' as SeverityLevel,
-      message: 'Touch event within element: Styled(View2)',
+      message: 'Touch event within element: Styled(View)',
       type: defaultProps.breadcrumbType,
     });
   });
 
-  it('maxComponentTreeSize', () => {
+  it('maxpathSize', () => {
     const { defaultProps } = TouchEventBoundary;
     const boundary = new TouchEventBoundary({
       ...defaultProps,
@@ -210,10 +210,104 @@ describe('TouchEventBoundary._onTouchStart', () => {
     expect(addBreadcrumb).toBeCalledWith({
       category: defaultProps.breadcrumbCategory,
       data: {
-        componentTree: ['Connect(View)', 'Styled(View)'],
+        path: [{ label: 'Connect(View)' }, { name: 'Styled(View)' }],
       },
       level: 'info' as SeverityLevel,
       message: 'Touch event within element: Connect(View)',
+      type: defaultProps.breadcrumbType,
+    });
+  });
+
+  // see https://docs.sentry.io/platforms/javascript/guides/react/features/component-names/
+  it('uses custom names provided by babel plugin', () => {
+    const { defaultProps } = TouchEventBoundary;
+    const boundary = new TouchEventBoundary(defaultProps);
+
+    const event = {
+      _targetInst: {
+        elementType: {
+          displayName: 'View',
+        },
+        memoizedProps: {
+          'data-sentry-component': 'Screen',
+          'data-sentry-element': 'AnimatedNativeScreen',
+          'data-sentry-source-file': 'screen.tsx',
+        },
+        return: {
+          elementType: {
+            displayName: 'Text',
+          },
+          return: {
+            memoizedProps: {
+              'custom-sentry-label-name': 'Connect(View)',
+              'data-sentry-component': 'MyView',
+              'data-sentry-element': 'unknown', // should be ignored
+              'data-sentry-source-file': 'myview.tsx',
+            },
+            return: {
+              elementType: {
+                displayName: 'Styled(View)',
+              },
+              return: {
+                memoizedProps: {
+                  'data-sentry-component': 'Happy',
+                  'data-sentry-element': 'View',
+                  'data-sentry-source-file': 'happyview.js',
+                },
+              },
+            },
+          },
+        },
+      },
+    };
+
+    // @ts-expect-error Calling private member
+    boundary._onTouchStart(event);
+
+    expect(addBreadcrumb).toBeCalledWith({
+      category: defaultProps.breadcrumbCategory,
+      data: {
+        path: [
+          { element: 'AnimatedNativeScreen', file: 'screen.tsx', name: 'Screen' },
+          { name: 'Text' },
+          { file: 'myview.tsx', name: 'MyView' },
+          { name: 'Styled(View)' },
+          { element: 'View', file: 'happyview.js', name: 'Happy' },
+        ],
+      },
+      level: 'info' as SeverityLevel,
+      message: 'Touch event within element: Screen (screen.tsx)',
+      type: defaultProps.breadcrumbType,
+    });
+  });
+
+  it('deduplicates', () => {
+    const { defaultProps } = TouchEventBoundary;
+    const boundary = new TouchEventBoundary(defaultProps);
+
+    const event = {
+      _targetInst: {
+        elementType: {
+          displayName: 'Text',
+        },
+        return: {
+          elementType: {
+            displayName: 'Text',
+          },
+        },
+      },
+    };
+
+    // @ts-expect-error Calling private member
+    boundary._onTouchStart(event);
+
+    expect(addBreadcrumb).toBeCalledWith({
+      category: defaultProps.breadcrumbCategory,
+      data: {
+        path: [{ name: 'Text' }],
+      },
+      level: 'info' as SeverityLevel,
+      message: 'Touch event within element: Text',
       type: defaultProps.breadcrumbType,
     });
   });